dopeblog/requirements_dev.txt
Django==2.2.8
django-taggit==1.2.0
pytz==2019.3
sqlparse==0.3.0
django-ckeditor==5.9.0
django-js-asset==1.2.2
